backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Redis
redis==5.0.1
aioredis==2.0.1

# Database
asyncpg==0.29.0

# LangChain and AI/ML
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
openai>=1.10.0,<2.0.0
faiss-cpu>=1.9.0
chromadb==0.4.18
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Security
cryptography==41.0.7
bcrypt==4.1.2

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Background tasks
celery==5.3.4

# Environment and configuration
python-decouple==3.8

# CORS
fastapi-cors==0.0.6

# Rate limiting
slowapi==0.1.9

# Health checks
healthcheck==1.3.3

# Production server
gunicorn==21.2.0
